[type: refactor] refactor pom
