+++ pom.xml
@@ -162,6 +162,7 @@
         <spring-cloud.version>3.1.2</spring-cloud.version>
         <log4j-1.2-api.vetsion>2.17.2</log4j-1.2-api.vetsion>
         <caffeine.version>2.9.3</caffeine.version>
+        <httpasyncclient.version>4.1.5</httpasyncclient.version>
     </properties>
 
     <dependencyManagement>
