+++ shenyu-client/pom.xml
@@ -33,7 +33,7 @@
         <module>shenyu-client-sofa</module>
         <module>shenyu-client-tars</module>
         <module>shenyu-client-grpc</module>
-        <module>shenyu-client-motan</module>
+    <!--    <module>shenyu-client-motan</module>-->
         <module>shenyu-client-websocket</module>
     </modules>
