+++ shenyu-plugin/shenyu-plugin-param-mapping/src/main/java/org/apache/shenyu/plugin/param/mapping/strategy/FormDataOperator.java
@@ -17,6 +17,7 @@
 
 package org.apache.shenyu.plugin.param.mapping.strategy;
 
+import com.google.gson.reflect.TypeToken;
 import com.jayway.jsonpath.DocumentContext;
 import org.apache.commons.collections4.CollectionUtils;
 import org.apache.shenyu.common.dto.convert.rule.impl.ParamMappingRuleHandle;
@@ -75,7 +76,7 @@ public class FormDataOperator implements Operator {
                     HttpHeaders httpHeaders = new HttpHeaders();
                     Charset charset = Objects.requireNonNull(headers.getContentType()).getCharset();
                     charset = charset == null ? StandardCharsets.UTF_8 : charset;
-                    LinkedMultiValueMap<String, String> modifyMap = GsonUtils.getInstance().toLinkedMultiValueMap(modify);
+                    LinkedMultiValueMap<String, String> modifyMap = toLinkedMultiValueMap(modify);
                     List<String> list = prepareParams(modifyMap, charset.name());
                     String content = String.join("&", list);
                     byte[] bodyBytes = content.getBytes(charset);
@@ -111,7 +112,19 @@ public class FormDataOperator implements Operator {
         }));
         return paramList;
     }
-
+    
+    /**
+     * To linked multiValue map.
+     *
+     * @param json the json
+     * @return the linked multiValue map
+     */
+    public LinkedMultiValueMap<String, String> toLinkedMultiValueMap(final String json) {
+        return GsonUtils.getGson().fromJson(json, new TypeToken<LinkedMultiValueMap<String, String>>() {
+        }.getType());
+    }
+    
+    
     static class ModifyServerHttpRequestDecorator extends ServerHttpRequestDecorator {
 
         private final HttpHeaders headers;
