+++ shenyu-sdk/shenyu-sdk-core/pom.xml
@@ -42,10 +42,5 @@
             <artifactId>shenyu-register-instance-core</artifactId>
             <version>${project.version}</version>
         </dependency>
-        <dependency>
-            <groupId>org.apache.shenyu</groupId>
-            <artifactId>shenyu-loadbalancer</artifactId>
-            <version>${project.version}</version>
-        </dependency>
     </dependencies>
 </project>
