+++ shenyu-sdk/shenyu-sdk-okhttp/pom.xml
@@ -35,6 +35,7 @@
         <dependency>
             <groupId>com.squareup.okhttp3</groupId>
             <artifactId>okhttp</artifactId>
+            <version>${okhttp.version}</version>
         </dependency>
     </dependencies>
 </project>
